Expanded stock list to multiple sizes, started work on new strategy of picking smallest compatible stock when new stock is needed

+++ Board.cs
@@ -9,7 +9,7 @@ namespace BoardCutter
     class Board
     {
         public double Price { get; set; }
-        public int Length { get; set; }
+        public double Length { get; set; }
 
         public double PricePerUnit 
         {
